pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "learning-path-assistant-hackathon-2025"
version = "1.0.0"
description = "A project for creating learning paths assistant for Hackathon 2025"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "az>=0.1.0.dev1",
    "asyncio>=3.4.3",
    "atlassian-python-api>=4.0.3",
    "azure-ai-projects>=1.0.0b10",
    "azure-core-tracing-opentelemetry>=1.0.0b12",
    "azure-identity>=1.21.0",
    "azure-monitor-opentelemetry>=1.6.8",
    "azure-search-documents>=11.5.2",
    "chainlit>=2.5.5",
    "httpx>=0.28.1",
    "opentelemetry-exporter-otlp>=1.31.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "pillow>=11.2.1",
    "pip>=25.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "semantic-kernel>=1.29.0",
    "terminal-colors>=3.0.0",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "pymongo>=4.12.1",
    "orjson>=3.10.18",
    "literalai>=0.1.201",
    "npx>=0.1.6",
    "asyncpg>=0.30.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth>=2.40.2",
    "google-api-python-client>=2.169.0",
    "dateparser>=1.2.1",
    "pytest>=8.4.1",
    "deepeval>=3.2.6",
    "trulens>=2.2.1",
    "trulens-eval>=2.2.1",
    "azure-ai-evaluation>=1.10.0",
    "trulens-providers-openai>=2.2.2",
    "openai>=1.100.1",
    "arize-phoenix-evals>=2.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "notebooks*", "frontend*"]

[tool.black]
line-length = 79
